+++ package.json
@@ -4,9 +4,11 @@
     "private": true,
     "dependencies": {
         "@testing-library/react": "^13.4.0",
+        "clsx": "^2.1.1",
         "cra-template-typescript": "1.2.0",
         "react": "^18.3.1",
         "react-dom": "^18.3.1",
+        "react-router-dom": "^7.1.1",
         "react-scripts": "5.0.1",
         "sass": "^1.83.1",
         "web-vitals": "^4.2.4"
