--- src/common/components/layouts/description/type.ts
@@ -1,3 +1,0 @@
-export type pageDescriptionProps = {
-  sources: string[];
-};
