+++ src/common/layouts/description/type.ts
@@ -1,0 +1,3 @@
+export interface pageDescriptionProps {
+    sources: string[];
+}
