+++ src/common/layouts/navigation/style.module.scss
@@ -1,0 +1,61 @@
+@import "@/styles/common.scss";
+
+.container {
+    position: relative;
+    left: 0;
+    width: 60px;
+    height: 100%;
+    border-right: 1px solid $system-black-01;
+    padding: 10px;
+}
+
+.upper {
+    border-bottom: 1px solid $system-grey-20;
+    @include flex-vertical(space-around);
+    padding: 10px 0;
+}
+
+.lower {
+    @include flex-vertical(space-around);
+    padding: 10px 0;
+}
+
+.menu {
+    position: relative;
+    width: 40px;
+    height: 40px;
+    border-radius: 2px;
+    cursor: pointer;
+    background-color: transparent;
+    @include flex-center;
+
+    &::before {
+        display: none;
+        content: "";
+        width: 2px;
+        height: 20px;
+        border-radius: 100px;
+        background-color: $brand-blue-50;
+        position: absolute;
+        left: 0;
+    }
+
+    a {
+        width: 100%;
+        height: 100%;
+    }
+}
+
+.selected {
+    background-color: $system-grey-10;
+
+    &::before {
+        display: inline-block;
+    }
+}
+
+.icon {
+    width: 20px;
+    height: 20px;
+    margin: 10px;
+}
